commands: move prev code to prev.rs
